requirements.txt
aiofiles==24.1.0
allure-pytest==2.14.3
allure-python-commons==2.14.3
alumnium==0.11.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
Appium-Python-Client==5.1.1
attrs==25.3.0
Authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.45
botocore==1.38.45
browser-use==0.4.0
bubus==1.2.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.22.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.29.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-anthropic==0.2.4
langchain-aws==0.2.27
langchain-core==0.3.66
langchain-deepseek==0.1.3
langchain-google-genai==2.1.5
langchain-ollama==0.3.3
langchain-openai==0.3.26
langchain-text-splitters==0.3.8
langsmith==0.4.2
markdown-it-py==3.0.0
markdownify==1.1.0
mdurl==0.1.2
mem0ai==0.1.112
numpy==1.26.4
ollama==0.5.1
openai==1.92.2
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
patchright==1.52.5
playwright==1.53.0
pluggy==1.6.0
portalocker==2.10.1
posthog==6.0.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
retry2==0.9.5
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
screeninfo==0.8.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uuid7==0.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
zstandard==0.23.0
